+++ lib/tasks/pull.ex
@@ -1,9 +1,9 @@
 defmodule Mix.Tasks.Dokku.Pull do
   @moduledoc """
-  Downloads your production database to your app root directory as `latest.dump`
-  and loads it into your development database. Useful for troubleshooting
-  production issues.
-"""
+    Downloads your production database to your app root directory as `latest.dump`
+    and loads it into your development database. Useful for troubleshooting
+    production issues.
+  """
   @shortdoc "Download production database and load into development database"
 
   alias ExDokku.Action
@@ -14,5 +14,7 @@ defmodule Mix.Tasks.Dokku.Pull do
     Action.start_ssh()
     Action.download_db("latest")
     Action.load_db("latest")
+
+    IO.puts("Pull complete")
   end
 end