+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.1 (2019-11-05)
 
 * Add mix dokku.reset_prod
+* Add mix dokku.restore_prod
 
 ## v0.1.0 (2019-10-17)